Fix warnings
- All warnings from Clang 10, GCC 9, and MSVC 19 are fixed or suppressed
- Change the type of parallel_for's start and end to size_t from long
- Type conversion warnings in MSVC are suppressed in parallel_for,
  tabulate, and delayed_sequence

+++ include/parlay/alloc.h
@@ -42,7 +42,7 @@ private:
   size_t num_small;
   size_t max_small;
   size_t max_size;
-  std::atomic<long> large_allocated{0};
+  std::atomic<size_t> large_allocated{0};
 
   std::unique_ptr<concurrent_stack<void*>[]> large_buckets;
   struct block_allocator *small_allocators;
