+++ include/parlay/internal/scheduler_plugins/omp.h
@@ -7,13 +7,13 @@ namespace parlay {
 
 // IWYU pragma: private, include "../../parallel.h"
 
-inline unsigned int num_workers() { return omp_get_max_threads(); }
-inline unsigned int worker_id() { return omp_get_thread_num(); }
+inline size_t num_workers() { return omp_get_max_threads(); }
+inline size_t worker_id() { return omp_get_thread_num(); }
 
 template <class F>
-inline void parallel_for(long start, long end, F f, long, bool) {
+inline void parallel_for(size_t start, size_t end, F f, size_t, bool) {
   _Pragma("omp parallel for")
-    for(long i=start; i<end; i++) f(i);
+    for(size_t i=start; i<end; i++) f(i);
 }
 
 bool in_par_do = false;
